% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_treatment_significance.R
\name{quick_sig_check}
\alias{quick_sig_check}
\title{Quick Significance Check}
\usage{
quick_sig_check(
  model_results_list,
  model_names = NULL,
  alpha = 0.05,
  primary_treatment = "treatment",
  focus_on_treatment_only = TRUE,
  use_pooled = TRUE
)
}
\description{
Quick Significance Check
}
